Layout the menu in a single line and adjust for small screens
This commit makes the menu more compact so more focus and space is given
to the content in the page. It lays out all the menu items across the
top of the screen (which is familiar to people).

It also adds some media-queries so that the menu works alright on
screens of different sizes.